Add Divide Two Integers solution
This one was rough. I don't have a way to debug C# on linux, and it's
really dampening my effectiveness in the language. The invalid test case
really didn't help matters at all, either. It took me far too long to
come up with a solution on this. Complaining aside, this was an
interesting exercise.